+++ packages/protocol/test/LengthFeeHook.t.sol
@@ -316,7 +316,7 @@ contract LengthFeeHookTest is Test, IERC721Receiver {
       author: user1,
       app: user2,
       channelId: channelId,
-      deadline: block.timestamp + 1 days,
+      deadline: uint64(block.timestamp + 1 days),
       parentId: bytes32(0)
     });
 
@@ -363,7 +363,7 @@ contract LengthFeeHookTest is Test, IERC721Receiver {
       author: user1,
       app: user2,
       channelId: channelId,
-      deadline: block.timestamp + 1 days,
+      deadline: uint64(block.timestamp + 1 days),
       parentId: bytes32(0)
     });
